move dup methods to abstract pemutilprovider class
